backend/pyproject.toml
[tool.poetry]
name = "fastapi-alembic-sqlmodel-async"
version = "0.1.0"
description = ""
authors = ["xavier_aguas <aguas.xavier2013a@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
alembic = "^1.7.6"
asyncpg = "^0.25.0"
fastapi = "^0.75.0"
sqlmodel = "^0.0.6"
uvicorn = "^0.17.5"
python-jose = "^3.3.0"
email-validator = "^1.1.3"
cryptography = "^36.0.1"
passlib = "^1.7.4"
SQLAlchemy-Utils = "^0.38.2"
black = "^22.1.0"
SQLAlchemy = "1.4.35"
python-multipart = "^0.0.5"
fastapi-pagination = {extras = ["sqlmodel"], version = "^0.9.1"}
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
aioredis = "^2.0.1"
google-cloud-dialogflow = "^2.15.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
